+++ db/CHANGELOG.md
@@ -3,6 +3,15 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [kevbot/db@next]
+
+### Added
+
+- Added `1.1.0_change.sql` script.
+  - This script adds two tables. One for tracking the database version. The other for a change log.
+- Added `mysql_manager.py`.
+  - This is a script to assist with the management of the mysql databases/servers.
+
 ## [kevbot/db@v1.0.0] - 2024-06-19
 
 ### Added